refactor: Move arrival classes from 'c2' package to 'stop' package
